found another error for parameters

+++ jarvis/modules/fun.py
@@ -60,13 +60,14 @@ async def mock(context):
             return
 
         reply_text = mocker(message)
-        if reply.sender.is_self:
-            try:
-                await reply.edit(reply_text)
-            except MessageNotModifiedError:
-                await context.edit("`This message is already mocked in the same way.`")
-                return
-            await context.delete()
+        if reply:
+            if reply.sender.is_self:
+                try:
+                    await reply.edit(reply_text)
+                except MessageNotModifiedError:
+                    await context.edit("`This message is already mocked in the same way.`")
+                    return
+                await context.delete()
         else:
             await context.edit(reply_text)
 command_help.update({
@@ -90,13 +91,14 @@ async def widen(context):
             return
 
         reply_text = str(message).translate(dict((i, i + 0xFEE0) for i in range(0x21, 0x7F)))
-        if reply.sender.is_self:
-            try:
-                await reply.edit(reply_text)
-            except MessageNotModifiedError:
-                await context.edit("`This message is already widened.`")
-                return
-            await context.delete()
+        if reply:
+            if reply.sender.is_self:
+                try:
+                    await reply.edit(reply_text)
+                except MessageNotModifiedError:
+                    await context.edit("`This message is already widened.`")
+                    return
+                await context.delete()
         else:
             await context.edit(reply_text)
 command_help.update({
@@ -123,12 +125,13 @@ async def fox(context):
 
         input_text = " ".join(message).lower()
         reply_text = corrupt(input_text)
-        if reply.sender.is_self:
-            try:
-                await reply.edit(reply_text)
-            except MessageNotModifiedError:
-                await context.edit("`Message have already been scratched.`")
-            await context.delete()
+        if reply:
+            if reply.sender.is_self:
+                try:
+                    await reply.edit(reply_text)
+                except MessageNotModifiedError:
+                    await context.edit("`Message have already been scratched.`")
+                await context.delete()
         else:
             await context.edit(reply_text)
 command_help.update({
